Refactor resistance module, replace magic numbers, improve naming
