--- crates/cfd-1d/src/resistance.rs
@@ -1,705 +1,0 @@
-//! Hydraulic resistance models for 1D CFD networks.
-//!
-//! This module provides comprehensive resistance modeling for various
-//! microfluidic components and flow conditions, including analytical
-//! solutions and empirical correlations.
-
-use cfd_core::error::{Error, Result};
-use cfd_core::fluid::Fluid;
-use nalgebra::RealField;
-use num_traits::cast::FromPrimitive;
-use serde::{Deserialize, Serialize};
-
-/// Trait for hydraulic resistance models
-pub trait ResistanceModel<T: RealField + Copy> {
-    /// Calculate hydraulic resistance [Pa·s/m³]
-    fn calculate_resistance(&self, fluid: &Fluid<T>, conditions: &FlowConditions<T>) -> Result<T>;
-
-    /// Get model name
-    fn model_name(&self) -> &str;
-
-    /// Get applicable Reynolds number range
-    fn reynolds_range(&self) -> (T, T);
-
-    /// Check if model is applicable for given conditions
-    fn is_applicable(&self, conditions: &FlowConditions<T>) -> bool {
-        let (re_min, re_max) = self.reynolds_range();
-        if let Some(re) = &conditions.reynolds_number {
-            *re >= re_min && *re <= re_max
-        } else {
-            true // Assume applicable if Re is unknown
-        }
-    }
-}
-
-/// Flow conditions for resistance calculations
-#[derive(Debug, Clone)]
-pub struct FlowConditions<T: RealField + Copy> {
-    /// Reynolds number
-    pub reynolds_number: Option<T>,
-    /// Flow velocity [m/s]
-    pub velocity: Option<T>,
-    /// Flow rate [m³/s]
-    pub flow_rate: Option<T>,
-    /// Temperature [K]
-    pub temperature: T,
-    /// Pressure [Pa]
-    pub pressure: T,
-}
-
-impl<T: RealField + Copy + FromPrimitive> FlowConditions<T> {
-    /// Create new flow conditions with default temperature and pressure
-    pub fn new(velocity: T) -> Self {
-        Self {
-            reynolds_number: None,
-            velocity: Some(velocity),
-            flow_rate: None,
-            temperature: T::from_f64(293.15)
-                .unwrap_or_else(|| T::from_f64(20.0).unwrap_or_else(T::one)), // 20°C
-            pressure: T::from_f64(101325.0)
-                .unwrap_or_else(|| T::from_f64(1.0).unwrap_or_else(T::one)), // 1 atm
-        }
-    }
-}
-
-/// Hagen-Poiseuille resistance model for circular channels
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct HagenPoiseuilleModel<T: RealField + Copy> {
-    /// Channel diameter [m]
-    pub diameter: T,
-    /// Channel length [m]
-    pub length: T,
-}
-
-impl<T: RealField + Copy> HagenPoiseuilleModel<T> {
-    /// Create a new Hagen-Poiseuille model
-    pub fn new(diameter: T, length: T) -> Self {
-        Self { diameter, length }
-    }
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> ResistanceModel<T>
-    for HagenPoiseuilleModel<T>
-{
-    fn calculate_resistance(&self, fluid: &Fluid<T>, conditions: &FlowConditions<T>) -> Result<T> {
-        let viscosity = fluid.dynamic_viscosity(conditions.temperature)?;
-        let pi = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero());
-        let onehundredtwentyeight = T::from_f64(128.0).unwrap_or_else(|| T::zero());
-
-        // R = (128 * μ * L) / (π * D^4)
-        let d4 =
-            num_traits::Float::powf(self.diameter, T::from_f64(4.0).unwrap_or_else(|| T::zero()));
-        let resistance = onehundredtwentyeight * viscosity * self.length / (pi * d4);
-
-        Ok(resistance)
-    }
-
-    fn model_name(&self) -> &str {
-        "Hagen-Poiseuille"
-    }
-
-    fn reynolds_range(&self) -> (T, T) {
-        (
-            T::zero(),
-            T::from_f64(cfd_core::constants::dimensionless::reynolds::PIPE_CRITICAL_LOWER)
-                .unwrap_or_else(|| T::zero()),
-        )
-    }
-}
-
-/// Rectangular channel resistance model with exact solution
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RectangularChannelModel<T: RealField + Copy> {
-    /// Channel width [m]
-    pub width: T,
-    /// Channel height [m]
-    pub height: T,
-    /// Channel length [m]
-    pub length: T,
-}
-
-impl<T: RealField + Copy> RectangularChannelModel<T> {
-    /// Create a new rectangular channel model
-    pub fn new(width: T, height: T, length: T) -> Self {
-        Self {
-            width,
-            height,
-            length,
-        }
-    }
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> ResistanceModel<T>
-    for RectangularChannelModel<T>
-{
-    fn calculate_resistance(&self, fluid: &Fluid<T>, conditions: &FlowConditions<T>) -> Result<T> {
-        let viscosity = fluid.dynamic_viscosity(conditions.temperature)?;
-        let aspect_ratio = self.width / self.height;
-
-        // Calculate friction factor using exact series solution
-        let f_re = self.calculate_friction_factor(aspect_ratio);
-
-        let area = self.width * self.height;
-        let dh = T::from_f64(4.0).unwrap_or_else(|| T::zero()) * area
-            / (T::from_f64(2.0).unwrap_or_else(|| T::zero()) * (self.width + self.height));
-
-        let resistance = f_re * viscosity * self.length / (area * dh * dh);
-
-        Ok(resistance)
-    }
-
-    fn model_name(&self) -> &str {
-        "Rectangular Channel (Exact)"
-    }
-
-    fn reynolds_range(&self) -> (T, T) {
-        (
-            T::zero(),
-            T::from_f64(cfd_core::constants::dimensionless::reynolds::PIPE_CRITICAL_LOWER)
-                .unwrap_or_else(|| T::zero()),
-        )
-    }
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> RectangularChannelModel<T> {
-    /// Calculate friction factor for rectangular channels
-    ///
-    /// # Accuracy Limitations
-    /// This implementation uses simplified approximations for numerical stability:
-    /// - Valid for aspect ratios between 0.1 and 10.0
-    /// - Accuracy decreases for extreme aspect ratios (< 0.1 or > 10.0)
-    /// - Maximum error ~5% for typical microfluidic geometries
-    /// - For high-precision applications, consider using the full series expansion
-    ///
-    /// # Applicable Range
-    /// - Reynolds number: Re < 2300 (laminar flow)
-    /// - Aspect ratio: 0.1 ≤ α ≤ 10.0 (recommended)
-    /// - Relative roughness: ε/Dh < 0.05
-    fn calculate_friction_factor(&self, aspect_ratio: T) -> T {
-        let alpha = if aspect_ratio >= T::one() {
-            aspect_ratio
-        } else {
-            T::one() / aspect_ratio
-        };
-
-        // Simplified friction factor calculation to avoid numerical issues
-        let twentyfour = T::from_f64(24.0).unwrap_or_else(|| T::zero());
-        let one = T::one();
-
-        if alpha >= one {
-            // Wide channel approximation (simplified)
-            // Based on Shah & London (1978) with numerical stabilization
-            let correction = one - T::from_f64(0.63).unwrap_or_else(|| T::zero()) / alpha;
-            twentyfour * RealField::max(correction, T::from_f64(0.1).unwrap_or_else(|| T::zero()))
-        // Ensure positive
-        } else {
-            // Tall channel approximation (simplified)
-            // Derived from reciprocal relationship with stabilization
-            let inv_alpha = one / alpha;
-            let base = T::from_f64(56.91).unwrap_or_else(|| T::zero());
-            base / RealField::max(inv_alpha, T::from_f64(0.1).unwrap_or_else(|| T::zero()))
-            // Ensure positive denominator
-        }
-    }
-}
-
-/// Darcy-Weisbach resistance model for turbulent flow
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DarcyWeisbachModel<T: RealField + Copy> {
-    /// Hydraulic diameter [m]
-    pub hydraulic_diameter: T,
-    /// Channel length [m]
-    pub length: T,
-    /// Surface roughness [m]
-    pub roughness: T,
-}
-
-impl<T: RealField + Copy> DarcyWeisbachModel<T> {
-    /// Create a new Darcy-Weisbach model
-    pub fn new(hydraulic_diameter: T, length: T, roughness: T) -> Self {
-        Self {
-            hydraulic_diameter,
-            length,
-            roughness,
-        }
-    }
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> ResistanceModel<T>
-    for DarcyWeisbachModel<T>
-{
-    fn calculate_resistance(&self, fluid: &Fluid<T>, conditions: &FlowConditions<T>) -> Result<T> {
-        let reynolds = conditions.reynolds_number.ok_or_else(|| {
-            Error::InvalidConfiguration(
-                "Reynolds number required for Darcy-Weisbach model".to_string(),
-            )
-        })?;
-
-        // Calculate friction factor using Colebrook-White equation (approximation)
-        let friction_factor = self.calculate_friction_factor(reynolds);
-
-        let area = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero())
-            * num_traits::Float::powf(
-                self.hydraulic_diameter,
-                T::from_f64(2.0).unwrap_or_else(|| T::zero()),
-            )
-            / T::from_f64(4.0).unwrap_or_else(|| T::zero());
-
-        // Convert Darcy friction factor to hydraulic resistance
-        let density = fluid.density;
-        let resistance = friction_factor * self.length * density
-            / (T::from_f64(2.0).unwrap_or_else(|| T::zero())
-                * area
-                * num_traits::Float::powf(
-                    self.hydraulic_diameter,
-                    T::from_f64(2.0).unwrap_or_else(|| T::zero()),
-                ));
-
-        Ok(resistance)
-    }
-
-    fn model_name(&self) -> &str {
-        "Darcy-Weisbach"
-    }
-
-    fn reynolds_range(&self) -> (T, T) {
-        (
-            T::from_f64(cfd_core::constants::dimensionless::reynolds::PIPE_CRITICAL_UPPER)
-                .unwrap_or_else(|| T::zero()),
-            T::from_f64(1e8).unwrap_or_else(|| T::zero()),
-        )
-    }
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> DarcyWeisbachModel<T> {
-    /// Calculate friction factor using Swamee-Jain approximation
-    fn calculate_friction_factor(&self, reynolds: T) -> T {
-        let relative_roughness = self.roughness / self.hydraulic_diameter;
-
-        // Swamee-Jain approximation to Colebrook-White equation
-        let term1 = relative_roughness / T::from_f64(3.7).unwrap_or_else(|| T::zero());
-        let term2 = T::from_f64(5.74).unwrap_or_else(|| T::zero())
-            / num_traits::Float::powf(reynolds, T::from_f64(0.9).unwrap_or_else(|| T::zero()));
-        let log_term = num_traits::Float::ln(term1 + term2);
-        T::from_f64(0.25).unwrap_or_else(|| T::zero())
-            / num_traits::Float::powf(log_term, T::from_f64(2.0).unwrap_or_else(|| T::zero()))
-    }
-}
-
-/// Entrance effects resistance model
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EntranceEffectsModel<T: RealField + Copy> {
-    /// Base resistance value
-    pub base_resistance: T,
-    /// Entrance length coefficient
-    pub entrance_coefficient: T,
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> ResistanceModel<T>
-    for EntranceEffectsModel<T>
-{
-    fn calculate_resistance(
-        &self,
-        _fluid: &Fluid<T>,
-        _conditions: &FlowConditions<T>,
-    ) -> Result<T> {
-        Ok(self.base_resistance * (T::one() + self.entrance_coefficient))
-    }
-
-    fn model_name(&self) -> &str {
-        "Entrance Effects"
-    }
-
-    fn reynolds_range(&self) -> (T, T) {
-        (T::zero(), T::from_f64(1e6).unwrap_or_else(|| T::zero()))
-    }
-}
-
-/// Methods for combining multiple resistance models
-#[derive(Debug, Clone, PartialEq)]
-pub enum CombinationMethod {
-    /// Add resistances in series
-    Series,
-    /// Add resistances in parallel
-    Parallel,
-    /// Custom weighted combination
-    Weighted(Vec<f64>),
-}
-
-/// Resistance model factory for creating standard models
-pub struct ResistanceModelFactory;
-
-impl ResistanceModelFactory {
-    /// Create Hagen-Poiseuille model for circular channel
-    pub fn hagen_poiseuille<T: RealField + Copy + FromPrimitive + Copy>(
-        diameter: T,
-        length: T,
-    ) -> HagenPoiseuilleModel<T> {
-        HagenPoiseuilleModel { diameter, length }
-    }
-
-    /// Create rectangular channel model
-    pub fn rectangular_channel<T: RealField + Copy + FromPrimitive + Copy>(
-        width: T,
-        height: T,
-        length: T,
-    ) -> RectangularChannelModel<T> {
-        RectangularChannelModel {
-            width,
-            height,
-            length,
-        }
-    }
-
-    /// Create Darcy-Weisbach model for turbulent flow
-    pub fn darcy_weisbach<T: RealField + Copy + FromPrimitive + Copy>(
-        hydraulic_diameter: T,
-        length: T,
-        roughness: T,
-    ) -> DarcyWeisbachModel<T> {
-        DarcyWeisbachModel {
-            hydraulic_diameter,
-            length,
-            roughness,
-        }
-    }
-}
-
-/// Simplified geometry enum for resistance model selection
-#[derive(Debug, Clone)]
-pub enum ChannelGeometry<T: RealField + Copy> {
-    /// Circular channel
-    Circular {
-        /// Diameter of the circular channel
-        diameter: T,
-        /// Length of the channel
-        length: T,
-    },
-    /// Rectangular channel
-    Rectangular {
-        /// Width of the rectangular channel
-        width: T,
-        /// Height of the rectangular channel
-        height: T,
-        /// Length of the channel
-        length: T,
-    },
-}
-
-/// Resistance calculator with model selection and validation
-pub struct ResistanceCalculator<T: RealField + Copy> {
-    _phantom: std::marker::PhantomData<T>,
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> ResistanceCalculator<T> {
-    /// Create new resistance calculator
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            _phantom: std::marker::PhantomData,
-        }
-    }
-
-    /// Calculate resistance with automatic model selection
-    pub fn calculate_auto(
-        &self,
-        geometry: &ChannelGeometry<T>,
-        fluid: &Fluid<T>,
-        conditions: &FlowConditions<T>,
-    ) -> Result<T> {
-        match geometry {
-            ChannelGeometry::Circular { diameter, length } => {
-                let model = HagenPoiseuilleModel {
-                    diameter: *diameter,
-                    length: *length,
-                };
-                model.calculate_resistance(fluid, conditions)
-            }
-            ChannelGeometry::Rectangular {
-                width,
-                height,
-                length,
-            } => {
-                let model = RectangularChannelModel {
-                    width: *width,
-                    height: *height,
-                    length: *length,
-                };
-                model.calculate_resistance(fluid, conditions)
-            }
-        }
-    }
-
-    /// Calculate resistance using Hagen-Poiseuille model
-    pub fn calculate_hagen_poiseuille(
-        &self,
-        diameter: T,
-        length: T,
-        fluid: &Fluid<T>,
-        conditions: &FlowConditions<T>,
-    ) -> Result<T> {
-        let model = HagenPoiseuilleModel { diameter, length };
-        model.calculate_resistance(fluid, conditions)
-    }
-
-    /// Calculate resistance using rectangular channel model
-    pub fn calculate_rectangular(
-        &self,
-        width: T,
-        height: T,
-        length: T,
-        fluid: &Fluid<T>,
-        conditions: &FlowConditions<T>,
-    ) -> Result<T> {
-        let model = RectangularChannelModel {
-            width,
-            height,
-            length,
-        };
-        model.calculate_resistance(fluid, conditions)
-    }
-
-    /// Calculate resistance using Darcy-Weisbach model
-    pub fn calculate_darcy_weisbach(
-        &self,
-        hydraulic_diameter: T,
-        length: T,
-        roughness: T,
-        fluid: &Fluid<T>,
-        conditions: &FlowConditions<T>,
-    ) -> Result<T> {
-        let model = DarcyWeisbachModel {
-            hydraulic_diameter,
-            length,
-            roughness,
-        };
-        model.calculate_resistance(fluid, conditions)
-    }
-}
-
-impl<T: RealField + Copy + FromPrimitive + num_traits::Float> Default for ResistanceCalculator<T> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use approx::assert_relative_eq;
-
-    #[test]
-    fn test_hagen_poiseuille() -> Result<()> {
-        let model = HagenPoiseuilleModel::new(100e-6, 0.001);
-        let fluid = cfd_core::fluid::Fluid::<f64>::water()?;
-        let conditions = FlowConditions::new(0.001);
-
-        let resistance = model.calculate_resistance(&fluid, &conditions)?;
-
-        // Verify resistance is positive
-        assert!(resistance > 0.0);
-
-        // Check pressure drop calculation
-        // For Hagen-Poiseuille flow: ΔP = R * Q
-        // where R is resistance and Q is flow rate
-        let flow_rate = 1e-6; // m³/s
-        let pressure_drop = resistance * flow_rate;
-        assert!(pressure_drop > 0.0);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_rectangular_channel() -> Result<()> {
-        let model = RectangularChannelModel::new(100e-6, 50e-6, 0.001);
-        let fluid = cfd_core::fluid::Fluid::<f64>::water()?;
-        let conditions = FlowConditions::new(0.001);
-
-        let resistance = model.calculate_resistance(&fluid, &conditions)?;
-
-        // Verify resistance is positive
-        assert!(resistance > 0.0);
-
-        // Verify hydraulic diameter calculation
-        // Calculate hydraulic diameter for rectangular channel
-        // D_h = 2 * w * h / (w + h)
-        let hydraulic_diameter = 2.0 * model.width * model.height / (model.width + model.height);
-        assert_relative_eq!(
-            hydraulic_diameter,
-            2.0 * 100e-6 * 50e-6 / (100e-6 + 50e-6),
-            epsilon = 1e-10
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_darcy_weisbach() -> Result<()> {
-        let diameter = 10e-3; // 10mm pipe for turbulent flow
-        let model = DarcyWeisbachModel::new(diameter, 0.001, 1e-6);
-
-        // Test with turbulent flow
-        let fluid = cfd_core::fluid::Fluid::<f64>::water()?;
-        let velocity = 1.0;
-        let density = 998.2; // Water density at 20°C
-        let viscosity = fluid.dynamic_viscosity(293.15)?;
-        let re = density * velocity * diameter / viscosity;
-
-        // Create flow conditions with Reynolds number
-        let mut conditions = FlowConditions::new(velocity);
-        conditions.reynolds_number = Some(re);
-
-        let resistance = model.calculate_resistance(&fluid, &conditions)?;
-
-        // Verify resistance is positive
-        assert!(resistance > 0.0);
-
-        // Verify Reynolds number is turbulent
-        assert!(re > 2300.0, "Reynolds number {} should be turbulent", re);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_flow_regime_detection() -> Result<()> {
-        let fluid = cfd_core::fluid::Fluid::<f64>::water()?;
-        let density = 998.2; // Water density at 20°C in kg/m³
-        let viscosity = fluid.dynamic_viscosity(293.15)?;
-
-        // Test laminar flow (small diameter, low velocity)
-        let laminar_diameter = 100e-6; // 100 micrometers
-        let laminar_velocity = 0.001;
-        let laminar_model = HagenPoiseuilleModel::new(laminar_diameter, 0.001);
-        let laminar_re = density * laminar_velocity * laminar_diameter / viscosity;
-        assert!(
-            laminar_re < 2300.0,
-            "Expected laminar flow, got Re = {}",
-            laminar_re
-        );
-
-        // Test turbulent flow (larger diameter, high velocity)
-        // For turbulent: Re = ρ*v*D/μ > 2300
-        // Need: D*v > 2300*μ/ρ = 2300*0.001/998.2 ≈ 0.0023
-        let turbulent_diameter = 10e-3; // 10 mm pipe
-        let turbulent_velocity = 1.0; // 1 m/s
-        let turbulent_re = density * turbulent_velocity * turbulent_diameter / viscosity;
-        assert!(
-            turbulent_re > 2300.0,
-            "Expected turbulent flow, got Re = {}",
-            turbulent_re
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_resistance_calculator() -> Result<()> {
-        let calculator = ResistanceCalculator::new();
-        let fluid = cfd_core::fluid::Fluid::<f64>::water()?;
-        let conditions = FlowConditions::new(0.001);
-
-        // Test Hagen-Poiseuille
-        let resistance =
-            calculator.calculate_hagen_poiseuille(100e-6, 0.001, &fluid, &conditions)?;
-        assert!(resistance > 0.0);
-
-        // Test rectangular
-        let resistance =
-            calculator.calculate_rectangular(100e-6, 50e-6, 0.001, &fluid, &conditions)?;
-        assert!(resistance > 0.0);
-
-        // Test Darcy-Weisbach with turbulent flow
-        let diameter = 10e-3; // 10mm pipe
-        let velocity = 1.0;
-        let density = 998.2;
-        let viscosity = fluid.dynamic_viscosity(293.15)?;
-        let re = density * velocity * diameter / viscosity;
-
-        let mut turbulent_conditions = FlowConditions::new(velocity);
-        turbulent_conditions.reynolds_number = Some(re);
-
-        let resistance = calculator.calculate_darcy_weisbach(
-            diameter,
-            0.001,
-            1e-6,
-            &fluid,
-            &turbulent_conditions,
-        )?;
-        assert!(resistance > 0.0);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_auto_selection() -> Result<()> {
-        let calculator = ResistanceCalculator::new();
-        let fluid = cfd_core::fluid::Fluid::<f64>::water()?;
-        let conditions = FlowConditions::new(0.001);
-
-        // Test with circular geometry
-        let circular_geom = ChannelGeometry::Circular {
-            diameter: 100e-6,
-            length: 0.001,
-        };
-
-        let resistance = calculator.calculate_auto(&circular_geom, &fluid, &conditions)?;
-        assert!(resistance > 0.0);
-
-        // Test with rectangular geometry
-        let rect_geom = ChannelGeometry::Rectangular {
-            width: 100e-6,
-            height: 50e-6,
-            length: 0.001,
-        };
-
-        let resistance = calculator.calculate_auto(&rect_geom, &fluid, &conditions)?;
-        assert!(resistance > 0.0);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_model_applicability() {
-        let hp_model = HagenPoiseuilleModel {
-            diameter: 100e-6,
-            length: 0.001,
-        };
-
-        let mut conditions = FlowConditions {
-            reynolds_number: Some(1000.0),
-            velocity: Some(0.1),
-            flow_rate: Some(1e-6),
-            temperature: 293.15,
-            pressure: 101325.0,
-        };
-
-        // Should be applicable for laminar flow
-        conditions.reynolds_number = Some(1000.0);
-        assert!(hp_model.is_applicable(&conditions));
-
-        // Should not be applicable for turbulent flow
-        conditions.reynolds_number = Some(5000.0);
-        assert!(!hp_model.is_applicable(&conditions));
-
-        // Should be applicable if Reynolds number is unknown
-        conditions.reynolds_number = None;
-        assert!(hp_model.is_applicable(&conditions));
-    }
-
-    #[test]
-    fn test_flow_conditions() {
-        let conditions = FlowConditions {
-            reynolds_number: Some(1000.0),
-            velocity: Some(0.1),
-            flow_rate: Some(1e-6),
-            temperature: 293.15,
-            pressure: 101325.0,
-        };
-
-        assert_eq!(conditions.reynolds_number, Some(1000.0));
-        assert_eq!(conditions.velocity, Some(0.1));
-        assert_eq!(conditions.flow_rate, Some(1e-6));
-        assert_eq!(conditions.temperature, 293.15);
-        assert_eq!(conditions.pressure, 101325.0);
-    }
-}
