+++ crates/cfd-validation/src/analytical/couette.rs
@@ -96,7 +96,9 @@ impl<T: RealField + Copy + FromPrimitive> AnalyticalSolution<T> for CouetteFlow<
     }
 
     fn domain_bounds(&self) -> [T; 6] {
-        let large = T::from_f64(1000.0).unwrap_or(T::from_f64(100.0).unwrap_or(T::one()));
+        const LARGE_DOMAIN_SIZE: f64 = 1000.0;
+        let large =
+            T::from_f64(LARGE_DOMAIN_SIZE).unwrap_or(T::from_f64(100.0).unwrap_or(T::one()));
         [
             T::zero(),
             large, // x: [0, L]
